Update Issue model to handle books
It's ugly but it works. Eventually I should add some sort of method on
the model to avoid having to do is_book checks all the time purely to
display the issue number. get_issue_name() or something.

(Made these changes directly to production because I need to do a ton
of work to get it running locally oops sorry)