Allow Markdown formatting in article subtitle
Cus the latest issue has _italics_. This may need to be updated in other
template files too but I think this is fine for now.

Note that the opengraph description will still have the formatting,
which is annoying but not the end of the world. At some point I should
do striptags or whatever it is.

Closes #108.